ncnn: simplified
