+++ project/ncnn/common/image/stb_image_impl.cpp
