+++ project/ncnn/common/postprocess/Mybuild
@@ -1,0 +1,16 @@
+package project.ncnn.common.postprocess
+
+@Build(stage=2)
+@BuildDepends(third_party.lib.ncnn.ncnn_public_headers)
+@BuildDepends(third_party.compiler_headers)
+@BuildDepends(embox.cxx.LibCxx)
+static module postprocess {
+
+    source
+        "dfl_decode.cpp",
+        "nms.cpp"
+
+    depends third_party.lib.ncnn.ncnn_public_headers
+    depends third_party.compiler_headers
+    depends embox.cxx.LibCxx
+}
