+++ project/ncnn/models/yolov8n/Mybuild
@@ -1,0 +1,15 @@
+package project.ncnn.models.yolov8n
+
+@Build(stage=2)
+@BuildDepends(third_party.lib.ncnn.libncnn)
+@BuildDepends(third_party.lib.ncnn.ncnn_public_headers)
+@BuildDepends(project.ncnn.common.postprocess.postprocess)
+module yolov8n_model {
+
+    source
+        "yolov8n_model.cpp"
+
+    depends third_party.lib.ncnn.libncnn
+    depends third_party.lib.ncnn.ncnn_public_headers
+    depends project.ncnn.common.postprocess.postprocess
+}
