+++ project/ncnn/models/yolov8n/assets/yolov8n_ncnn_model/metadata.yaml
