+++ project/ncnn/models/yolov8n/assets/yolov8n_ncnn_model/yolov8n.id.h
